requirements.txt
pydantic-ai==0.0.43
httpx>=0.27.0
logfire>=0.8.0
openai>=1.0.0
streamlit>=1.30.0
streamlit-autorefresh>=1.0.1
google-auth-oauthlib>=0.4.6
google-auth-httplib2
google-api-python-client>=2.35.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.3
langchain>=0.1.0
python-dotenv>=1.0.0
supabase>=2.0.0
pydantic>=2.0.0
aiohttp>=3.8.3
async-timeout>=4.0.2
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.3
pdf2image>=1.16.3
PyPDF2>=3.0.1
python-docx>=0.8.11
pytesseract>=0.3.10
aiolimiter>=1.1.0
markdown>=3.4.4
beautifulsoup4>=4.12.2
pytest>=7.4.0
pytest-asyncio>=0.21.1 